beginning to move CLI into one file..and methods in another
