+++ bin/run.rb
@@ -2,7 +2,6 @@ require_relative '../config/environment'
 require_all 'lib'
 
 
-
 #
 # search_by_artist("magritte")
 #
